\name{spleen-package}
\alias{spleen-package}
\alias{spleen}
\docType{package}
\title{
\packageTitle{spleen}
}
\description{
\packageDescription{spleen}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{spleen}
\packageIndices{spleen}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{spleen}

Maintainer: \packageMaintainer{spleen}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
